chore: rename directory names as temporary name for Input/Table
