+++ client/components/page/Admin/Plan/PlanDetail.tsx
@@ -1,4 +1,4 @@
-import { ReactHookFormInput, TInputField } from '@/components/shared/Input';
+import { ReactHookFormInput, TInputField } from '@/components/shared/Inputtmp';
 import { Flex, useColorModeValue, VStack, Button } from '@chakra-ui/react';
 import { PlanDetailNextPageProps } from 'pages/admin/plan/[planId]';
 import { useForm } from 'react-hook-form';
