+++ client/components/page/Admin/Promotion/PromotionDetail.tsx
@@ -1,4 +1,4 @@
-import { ReactHookFormInput, TInputField } from '@/components/shared/Input';
+import { ReactHookFormInput, TInputField } from '@/components/shared/Inputtmp';
 import { Flex, useColorModeValue, VStack, Button } from '@chakra-ui/react';
 import { PromotionDetailNextPageProps } from 'pages/admin/promotion/[promotionId]';
 import { useForm } from 'react-hook-form';
